+++ action_app/bin/main.dart
@@ -4,13 +4,18 @@ import 'package:action_app/action_app.dart';
 import 'package:actions_toolkit_dart/core.dart';
 import 'package:dart_code_metrics/config.dart';
 import 'package:dart_code_metrics/lint_analyzer.dart';
+import 'package:dart_code_metrics/unused_files_analyzer.dart';
 
 Future<void> main() async {
   final workflowUtils =
       GitHubWorkflowUtils(environmentVariables: Platform.environment);
 
   final arguments = Arguments(workflowUtils);
-  final reporter = await AnalyzeReporter.create(
+  final analyzeReporter = await AnalyzeReporter.create(
+    workflowUtils: workflowUtils,
+    arguments: arguments,
+  );
+  final unusedFilesFileReport = await UnusedFilesReporter.create(
     workflowUtils: workflowUtils,
     arguments: arguments,
   );
@@ -25,27 +30,43 @@ Future<void> main() async {
 
     _getTheTargetPackagesDependencies(pubspecUtils, rootFolder);
 
-    await reporter.run();
+    await analyzeReporter.run();
+    await unusedFilesFileReport.run();
 
     startGroup(name: 'Running Dart Code Metrics');
 
     final foldersToAnalyze = arguments.folders;
     final options = await analysisOptionsFromFilePath(rootFolder);
-    final config = LintConfig.fromAnalysisOptions(options);
+    final lintConfig = LintConfig.fromAnalysisOptions(options);
 
     final lintAnalyzerReport = await const LintAnalyzer()
-        .runCliAnalysis(foldersToAnalyze, rootFolder, config);
+        .runCliAnalysis(foldersToAnalyze, rootFolder, lintConfig);
 
-    await reporter.complete(
+    await analyzeReporter.complete(
       pubspecUtils.packageName,
       foldersToAnalyze,
       lintAnalyzerReport,
     );
 
+    final unusedFilesConfig = UnusedFilesConfig.fromAnalysisOptions(options);
+
+    final unusedFilesReport = await const UnusedFilesAnalyzer().runCliAnalysis(
+      foldersToAnalyze,
+      rootFolder,
+      unusedFilesConfig,
+    );
+
+    await unusedFilesFileReport.complete(
+      pubspecUtils.packageName,
+      foldersToAnalyze,
+      unusedFilesReport,
+    );
+
     endGroup();
   } on Exception catch (cause) {
     try {
-      await reporter.cancel(cause: cause);
+      await analyzeReporter.cancel(cause: cause);
+      await unusedFilesFileReport.cancel(cause: cause);
       // ignore: avoid_catches_without_on_clauses
     } catch (exception, stackTrace) {
       error(message: '$exception\n$stackTrace');