+++ action_app/lib/action_app.dart
@@ -3,3 +3,4 @@ export 'package:action_app/src/arguments.dart';
 export 'package:action_app/src/git_utils.dart';
 export 'package:action_app/src/github_workflow_utils.dart';
 export 'package:action_app/src/pubspec_utils.dart';
+export 'package:action_app/src/unused_files_reporter.dart';