+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Added `check-unused-files` report
+
 ## 1.0.1
 
 * Added support **GitHub Personal access token**.